% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{remove_constant_columns}
\alias{remove_constant_columns}
\title{Remove constant columns in data}
\usage{
remove_constant_columns(X)
}
\arguments{
\item{X}{A data frame.}
}
\value{
A data frame where constant columns have been removed.
}
\description{
Given some data X, removes all columns in the data that are a
constant value (while ignoring NAs).
}
